Add estimate of years to crack with -v
My estimate takes into account sorting alphabetically.

Note that I believe Monroe got his calculations wrong for the number
of years to crack a passphrase with 44 bits of entropy. I think he
mistakenly calculated the *maximium* amount of time to crack instead
of the average.

2^44 / 1000 / 60 / 60 / 24 / 365 is indeed a little over 550 years, as
he mentioned in the XKCD 936. However, that is only true if you one
had to try every single possibility. On average, the attack will
succeed in half that time.