Correct entropy calculation for chbs.net
I had mistakenly though CorrectHorseBatteryStaple.net had a file of
15627 words, but it was actually 15627 *characters*. The true count is
2284 words, which is just about perfect given XKCD 936's assumption
that the list would be 2048 words.

(Despite the similarity in size, this is not the word list that XKCD
936 used as it is missing the words "battery" and "staple". It appears
to instead be John Baumann's 1995 update of the 1953 General Service
List, which contains words that are not necessarily the most frequent,
but are the most important for a person learning English to know.
http://jbauman.com/aboutgsl.html ).

Bits of Entropy = 4*log2(2284)+log2(10) = 47.95